Separate into 2 pages, email page & info page
