+++ walk-in-interview/src/main/webapp/account/create-account/business/account-info/index.html
@@ -1,0 +1,24 @@
+<!DOCTYPE html>
+<html>
+  <head>
+    <meta charset="UTF-8">
+    <link rel="stylesheet" href="../../../../style.css">
+  </head>
+
+  <body>
+    <div class="header-container">
+      <button id="back"></button>
+      <input id="submit">
+    </div>
+
+    <div class="content-container">
+      <label id="name-label" for="name"></label>
+      <input id="name" name="name">
+    </div>
+
+    <div class="error-message" id="error-message"></div>
+  </body>
+
+  <script type="module" src="./script.js"></script>
+</html>
+
