+++ walk-in-interview/src/main/webapp/account/create-account/business/index.html
@@ -6,19 +6,33 @@
   </head>
 
   <body>
-    <div class="header-container">
+    <div class="log-in-section">
+      <ul>
+        <li id="account">
+          <label for="account-input"></label>
+          <input id="account-input">
+        </li>
+  
+        <li id="password">
+          <label for="password-input"></label>
+          <input id="password-input">
+        </li>
+      </ul>
+
+      <br>
+
       <button id="back"></button>
-      <input id="submit">
-    </div>
+      <button id="submit"></button>
 
-    <div class="content-container">
-      <label id="name-label" for="name"></label>
-      <input id="name" name="name">
+      <div class="error-message" id="error-message"></div>
     </div>
-
-    <div class="error-message" id="error-message"></div>
   </body>
 
   <script type="module" src="./script.js"></script>
-</html>
 
+  <!-- TODO(issue/87): move to separate config file -->
+  <script src="https://www.gstatic.com/firebasejs/7.17.1/firebase-app.js"></script>
+  <script src="https://www.gstatic.com/firebasejs/7.17.1/firebase-auth.js"></script>
+  <script src="https://cdn.firebase.com/libs/firebaseui/3.5.2/firebaseui.js"></script>
+  <link type="text/css" rel="stylesheet" href="https://cdn.firebase.com/libs/firebaseui/3.5.2/firebaseui.css" />
+</html>
